requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
django-filter==23.2
djangorestframework==3.14.0
gunicorn==20.1.0
Markdown==3.4.3
pytz==2023.3
Serializer==0.2.1
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.6.3
whitenoise==6.5.0
